Move video back to master, move video into main

+++ Find/MainController/RealmFiles/RealmComponent.swift
@@ -13,10 +13,10 @@ class RealmComponent: Object {
     @objc dynamic var title: String = ""
     @objc dynamic var done: Bool = false
     @objc dynamic var dateCreated: Date?
-    @objc dynamic var x: CGFloat = 0
-    @objc dynamic var y: CGFloat = 0
-    @objc dynamic var height: CGFloat = 0
-    @objc dynamic var width: CGFloat = 0
+    @objc dynamic var x: Float = 0
+    @objc dynamic var y: Float = 0
+    @objc dynamic var height: Float = 0
+    @objc dynamic var width: Float = 0
     @objc dynamic var text = ""
 //    var parentCategory = LinkingObjects(fromType: Photo.self, property: "items")
     @objc dynamic var photo: RealmPhoto!
